comic demo added NavigationDrawer
