+++ example/lib/main.dart
@@ -9,7 +9,7 @@ import 'package:example/screens/checkbox/checkbox.screen.dart';
 import 'package:example/screens/checkbox_list_tile/checkbox_list_tile.screen.dart';
 import 'package:example/screens/chip/chip.screen.dart';
 import 'package:example/screens/color_scheme/color_scheme.screen.dart';
-import 'package:example/screens/demo/comic.theme.screen.dart';
+import 'package:example/screens/demo/comic/comic.theme.screen.dart';
 import 'package:example/screens/demo/sleek.theme.screen.dart';
 import 'package:example/screens/divider/divider.screen.dart';
 import 'package:example/screens/floating_action_button/floating_action_button.screen.dart';
