+++ example/lib/screens/demo/comic/chat.tab.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ChatTab extends StatelessWidget {
+  const ChatTab({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: const Text("Chat Tab"),
+    );
+  }
+}
