+++ example/lib/screens/demo/comic/profile.tab.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ProfileTab extends StatelessWidget {
+  const ProfileTab({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: const Text('Profile Tab'),
+    );
+  }
+}
