+++ example/lib/screens/demo/comic/settings.tab.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class SettingsTab extends StatelessWidget {
+  const SettingsTab({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: const Text("Settings"),
+    );
+  }
+}
